testutil: ssntp test certificates cleanup and test coverage
The role to cert lookup makes more sense in testutil/ssntp_test_certs.go
and that file needs test coverage.
