+++ testutil/testutil.go
@@ -16,10 +16,6 @@
 
 package testutil
 
-import (
-	"github.com/01org/ciao/ssntp"
-)
-
 // CmdResult is a common result structure for tests spanning between
 // controller client, scheduler server, and the various (eg: Agent,
 // NetAgent, CNCIAgent) agent roles.
@@ -29,27 +25,4 @@ type CmdResult struct {
 	NodeUUID     string
 	TenantUUID   string
 	CNCI         bool
-}
-
-// RoleToTestCert returns a string containing the testutil certificate
-// matching the specified ssntp.Role
-func RoleToTestCert(role ssntp.Role) string {
-	switch role {
-	case ssntp.SCHEDULER:
-		return TestCertScheduler
-	case ssntp.SERVER:
-		return TestCertServer
-	case ssntp.AGENT:
-		return TestCertAgent
-	case ssntp.Controller:
-		return TestCertController
-	case ssntp.CNCIAGENT:
-		return TestCertCNCIAgent
-	case ssntp.NETAGENT:
-		return TestCertNetAgent
-	case ssntp.AGENT | ssntp.NETAGENT:
-		return TestCertAgentNetAgent
-	}
-
-	return TestCertUnknown
 }
